+++ NEWS.md
@@ -1,3 +1,7 @@
+# nfl4th (development version)
+
+* The internal cache is now aware of changed model formats and forces a model update for compatibility with xgboost (>= v3).
+
 # nfl4th 1.0.5
 
 * Stop letting roof break things